[#421] Allow regex in user claim plugin to modify claim
This commit allows for two configuration options, `match_regex`
and `replace_format` to modify a matched claim according to the
regular expression given in the new options. Its intended use
is to help those unable to control how a claim is returned from
an OpenID Provider, modify the returned claim to a desired result.